+++ src/main/java/io/github/axolotlclient/modules/hud/gui/hud/vanilla/CrosshairHud.java
@@ -77,7 +77,7 @@ public class CrosshairHud extends AbstractHudEntry implements DynamicallyPositio
                     new int[] { 0, 0, 0, 0, 0, 0, 0, -1, 0, 0, 0, 0, 0, 0, 0 },
                     new int[] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
                     new int[] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 },
-                    new int[] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 }, });
+                    new int[] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 }, }, true);
 
     public CrosshairHud() {
         super(15, 15);